finalvers

+++ Interfaces/IUser.cs
@@ -5,6 +5,6 @@ namespace MoviesTelegramBot.Interfaces;
 public interface IUser
 {
     Task AddMovie(long userId, string movie);
-    Task RemoveMovie(long userId, long movieId);
+    Task RemoveMovie(long userId, int movieId);
     Task<List<string>> GetAllMoviesFromUser(long userId);
 }
